Started adding int128, int256 tests
